requirements.txt
differnet-zerobox
# scikit-learn>=0.22
# scipy>=1.3.2
# numpy>=1.17.4
# torch==1.8.1
# torchvision==0.9.1
# matplotlib>=3.0.3
# tqdm>=4.59.2
# opencv-python>=4.5.1